# docs/swagger/swagger.yaml
basePath: /api/v2
definitions:
...
  common.JSONResult:
    properties:
      data:
        description: 数据
      extra:
        description: 数据
      msg:
        description: 消息
        type: string
      success:
        description: 是否成功
        type: boolean
    type: object
...
  entity.CookedComment:
    properties:
      badge_color:
        type: string
      badge_name:
        type: string
      content:
        type: string
      content_marked:
        type: string
      date:
        type: string
      email_encrypted:
        type: string
      id:
        type: integer
      ip_region:
        type: string
      is_allow_reply:
        type: boolean
      is_collapsed:
        type: boolean
      is_pending:
        type: boolean
      is_pinned:
        type: boolean
      link:
        type: string
      nick:
        type: string
      page_key:
        type: string
      page_url:
        type: string
      rid:
        type: integer
      site_name:
        type: string
      ua:
        type: string
      user_id:
        type: integer
      visible:
        type: boolean
      vote_down:
        type: integer
      vote_up:
        type: integer
    type: object
...
  entity.CookedPage:
    properties:
      admin_only:
        type: boolean
      id:
        type: integer
      key:
        type: string
      pv:
        type: integer
      site_name:
        type: string
      title:
        type: string
      url:
        type: string
      vote_down:
        type: integer
      vote_up:
        type: integer
    type: object
  entity.CookedSite:
    properties:
      first_url:
        type: string
      id:
        type: integer
      name:
        type: string
      urls:
        items:
          type: string
        type: array
      urls_raw:
        type: string
    type: object
...
  handler.ParamsAdminPageEdit:
    properties:
      admin_only:
        description: Edit page admin_only option
        type: boolean
      key:
        description: Edit page key
        type: string
      site_name:
        description: The site name of your content scope
        type: string
      title:
        description: Edit page title
        type: string
    type: object
  handler.ParamsAdminPageFetch:
    properties:
      get_status:
        description: If true, only get the status of the current task status
        type: boolean
      site_name:
        description: The site name of your content scope
        type: string
    type: object
...
  handler.ParamsAdminSettingSave:
    properties:
      data:
        description: The content of the config file in YAML format
        type: string
    required:
    - data
    type: object
...
  handler.ParamsAdminSiteEdit:
    properties:
      name:
        description: Edit site name
        type: string
      urls:
        description: Edit site urls
        type: string
    type: object
...
  handler.ParamsCommentEdit:
    properties:
      content:
        description: The comment content
        type: string
      email:
        description: The comment email
        type: string
      ip:
        description: The comment ip
        type: string
      is_collapsed:
        description: The comment is_collapsed
        type: boolean
      is_pending:
        description: The comment is_pending
        type: boolean
      is_pinned:
        description: The comment is_pinned
        type: boolean
      link:
        description: The comment link
        type: string
      nick:
        description: The comment nick
        type: string
      page_key:
        description: The comment page_key
        type: string
      rid:
        description: The comment rid
        type: string
      site_name:
        description: The site name of your content scope
        type: string
      ua:
        description: The comment ua
        type: string
    type: object
...
  handler.ResponseAdminPageEdit:
    properties:
      page:
        $ref: '#/definitions/entity.CookedPage'
    type: object
...
  handler.ResponseAdminSettingGet:
    properties:
      custom:
        type: string
      template:
        type: string
    type: object
...
  handler.ResponseAdminSiteEdit:
    properties:
      site:
        $ref: '#/definitions/entity.CookedSite'
    type: object
...
  handler.ResponseCommentEdit:
    properties:
      comment:
        $ref: '#/definitions/entity.CookedComment'
    type: object
...
info:
  contact:
    email: artalkjs@gmail.com
    name: API Support
    url: https://artalk.js.org
  description: Artalk is a modern comment system based on Golang.
  license:
    name: MIT
    url: https://github.com/ArtalkJS/Artalk/blob/master/LICENSE
  title: Artalk API
  version: "2.0"
paths:
...
  /comments/{id}:
    delete:
      description: Delete a specific comment
      parameters:
      - description: The comment ID you want to delete
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JSONResult'
      security:
      - ApiKeyAuth: []
      summary: Delete Comment
      tags:
      - Comment
    put:
      consumes:
      - application/json
      description: Edit a specific comment
      parameters:
      - description: The comment ID you want to edit
        in: path
        name: id
        required: true
        type: integer
      - description: The comment data
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/handler.ParamsCommentEdit'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/handler.ResponseCommentEdit'
              type: object
      security:
      - ApiKeyAuth: []
      summary: Edit Comment
      tags:
      - Comment
...
  /pages/{id}:
    delete:
      description: Delete a specific page
      parameters:
      - description: The page ID you want to delete
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JSONResult'
      security:
      - ApiKeyAuth: []
      summary: Delete Page
      tags:
      - Page
    put:
      consumes:
      - application/json
      description: Edit a specific page
      parameters:
      - description: The page ID you want to edit
        in: path
        name: id
        required: true
        type: integer
      - description: The page data
        in: body
        name: page
        required: true
        schema:
          $ref: '#/definitions/handler.ParamsAdminPageEdit'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/handler.ResponseAdminPageEdit'
              type: object
      security:
      - ApiKeyAuth: []
      summary: Edit Page
      tags:
      - Page
  /pages/{id}/fetch:
    post:
      consumes:
      - application/json
      description: Fetch the data of a specific page
      parameters:
      - description: The page ID you want to fetch
        in: path
        name: id
        required: true
        type: integer
      - description: The options
        in: body
        name: options
        required: true
        schema:
          $ref: '#/definitions/handler.ParamsAdminPageFetch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JSONResult'
      security:
      - ApiKeyAuth: []
      summary: Fetch Page Data
      tags:
      - Page
...
  /settings:
    get:
      description: Get settings from app config file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/handler.ResponseAdminSettingGet'
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get Settings
      tags:
      - System
    post:
      consumes:
      - application/json
      description: Save settings to app config file
      parameters:
      - description: The settings data
        in: body
        name: settings
        required: true
        schema:
          $ref: '#/definitions/handler.ParamsAdminSettingSave'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JSONResult'
      security:
      - ApiKeyAuth: []
      summary: Save Settings
      tags:
      - System
  /settings/template/{locale}:
    get:
      description: Get config templates in different languages for rendering the settings
        page in the frontend
      parameters:
      - description: The locale of the settings template you want to get
        in: path
        name: locale
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get Settings Template
      tags:
      - System
...
  /sites/{id}:
    delete:
      description: Delete a specific site
      parameters:
      - description: The site ID you want to delete
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JSONResult'
      security:
      - ApiKeyAuth: []
      summary: Delete Site
      tags:
      - Site
    put:
      consumes:
      - application/json
      description: Edit a specific site
      parameters:
      - description: The site ID you want to edit
        in: path
        name: id
        required: true
        type: string
      - description: The site data
        in: body
        name: site
        required: true
        schema:
          $ref: '#/definitions/handler.ParamsAdminSiteEdit'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/handler.ResponseAdminSiteEdit'
              type: object
      security:
      - ApiKeyAuth: []
      summary: Edit Site
      tags:
      - Site
...
securityDefinitions:
  ApiKeyAuth:
    description: '"Type ''Bearer TOKEN'' to correctly set the API Key"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"